[Impeller] improved documentation of the gaussian blur.
