+++ impeller/entity/BUILD.gn
@@ -158,10 +158,10 @@ impeller_component("entity") {
     "contents/filters/color_filter_contents.h",
     "contents/filters/color_matrix_filter_contents.cc",
     "contents/filters/color_matrix_filter_contents.h",
+    "contents/filters/directional_gaussian_blur_filter_contents.cc",
+    "contents/filters/directional_gaussian_blur_filter_contents.h",
     "contents/filters/filter_contents.cc",
     "contents/filters/filter_contents.h",
-    "contents/filters/gaussian_blur_filter_contents.cc",
-    "contents/filters/gaussian_blur_filter_contents.h",
     "contents/filters/inputs/contents_filter_input.cc",
     "contents/filters/inputs/contents_filter_input.h",
     "contents/filters/inputs/filter_contents_filter_input.cc",
